+++ zinit-autoload.zsh
@@ -3303,16 +3303,18 @@ EOF
 # It's an attempt to plugin only this one function into `zinit' function
 # defined in zinit.zsh, to not make this file longer than it's needed.
 .zinit-module() {
+    local module_dir="${ZINIT[HOME_DIR]}/module"
     if [[ "$1" = "build" ]]; then
         .zinit-build-module "${@[2,-1]}"
     elif [[ "$1" = "info" ]]; then
         if [[ "$2" = "--link" ]]; then
               builtin print -r "You can copy the error messages and submit"
-              builtin print -r "error-report at: https://github.com/zdharma-continuum/zinit/issues"
+              builtin print -r "error-report at: https://github.com/zdharma-continuum/zinit-module/issues"
         else
             builtin print -r "To load the module, add following 2 lines to .zshrc, at top:"
-            builtin print -r "    module_path+=( \"${ZINIT[BIN_DIR]}/zmodules/Src\" )"
-            builtin print -r "    zmodload zdharma/zplugin"
+            # TODO: Update when we move the C code source to the main dir of the repo
+            builtin print -r "    module_path+=( \"${module_dir}/zmodules/Src\" )"
+            builtin print -r "    zmodload zdharma/zinit"
             builtin print -r ""
             builtin print -r "After loading, use command \`zpmod' to communicate with the module."
             builtin print -r "See \`zpmod -h' for more information."
@@ -3335,9 +3337,14 @@ EOF
 .zinit-build-module() {
     setopt localoptions localtraps
     trap 'return 1' INT TERM
-    ( builtin cd -q "${ZINIT[BIN_DIR]}"/zmodules
-      +zinit-message "{pname}== Building module zdharma/zplugin, running: make clean, then ./configure and then make =={rst}"
-      +zinit-message "{pname}== The module sources are located at: "${ZINIT[BIN_DIR]}"/zmodules =={rst}"
+    local module_dir="${ZINIT[HOME_DIR]}/module"  # TODO: Make this ZINIT[MODULE_DIR] ie configurable
+    command git clone "https://github.com/zdharma-continuum/zinit-module.git" "${module_dir}" || {
+        builtin print "${ZINIT[col-error]}Failed to clone module repo${ZINIT[col-rst]}"
+        return 1
+    }
+    ( builtin cd -q "${module_dir}/zmodules"  # TODO: Move the source of the module to the root of the repo
+      +zinit-message "{pname}== Building module zdharma-continuum/zinit-module, running: make clean, then ./configure and then make =={rst}"
+      +zinit-message "{pname}== The module sources are located at: "${module_path}" =={rst}"
       if [[ -f Makefile ]] {
           if [[ "$1" = "--clean" ]] {
               noglob +zinit-message {p}-- make distclean --{rst}
@@ -3360,7 +3367,7 @@ EOF
               .zinit-module info --link
           }
       }
-      builtin print $EPOCHSECONDS >! "${ZINIT[BIN_DIR]}"/zmodules/COMPILED_AT
+      builtin print $EPOCHSECONDS >! "${module_path}/COMPILED_AT"
     )
 }
 # ]]]